Supporting retrieval of an arbitrary number of columns when using `get`.
Retrieval of columns should respect the char limit of the url, if the
length of the columns string will exceed that we need to chunk it up
in order to return the requested columns.

HBase was built to support thousands of columns and this change is necessary
for the JSONRest module to be able to fully utilize the powers of
HBase column capacity.